+++ app/src/main/java/com/main/proyek_salez/data/repository/SalezRepositories.kt
@@ -31,7 +31,7 @@ class SalezRepository @Inject constructor(
         return foodItemDao.searchFoodItems(name)
             .map { entities -> entities.map { it.toFoodItem() } }
     }
-    suspend fun getFoodItemById(id: Long): FoodItem? {
+    suspend fun getFoodItemById(id: Int): FoodItem? {
         return foodItemDao.getFoodItemById(id)?.toFoodItem()
     }
